Refied fuctioality & resposiveness
Refactored my context process (Fetch.js) to call the fetch function upon loading.

Added conditionals to handle errors and improve user experience, more specifically if a user does not provide a location and pushes a button to select the type of pet.

The pet type buttons call back the fetch function as well, but the page would break if there was no location indicated. It doesn't anymore.